requirements.txt
fastapi==0.115.12
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.61
langchain-ollama==0.2.3
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
uvicorn==0.34.2
websockets==15.0.1
python-dotenv==1.1.0
numpy==1.26.4
ollama==0.5.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyodbc==5.2.0